Rename CAN -> CAN bis to support duplicate merge
